Fix delete character, can only delete characters that the current user have

+++ Services/CharacterService/CharacterService.cs
@@ -47,11 +47,21 @@ namespace dotnet_rpg.Services.CharacterService
 
       try
       {
-        var character = await _context.Characters.FirstAsync(c => c.Id == id);
-        _context.Characters.Remove(character);
-
-        await _context.SaveChangesAsync();
-        response.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+        var character = await _context.Characters
+          .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+        if(character != null)
+        {
+          _context.Characters.Remove(character);
+          await _context.SaveChangesAsync();
+          response.Data = _context.Characters
+            .Where(c => c.User.Id == GetUserId())
+            .Select(c => _mapper.Map<GetCharacterDto>(c))
+            .ToList();
+        }
+        else {
+          response.Success = false;
+          response.Message = "Character not found.";
+        }
       }
       catch (Exception ex)
       {
